user_profile_service/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic
email-validator==2.2.0
python-dotenv
cryptography
python-jose[cryptography]
passlib[bcrypt]
